identity: keep the setfsuid wrappers Linux-only so darwin builds compile
